Move packages to com.victoriametrics.client
