+++ src/main/java/com/victoriametrics/client/metrics/Counter.java
@@ -1,4 +1,4 @@
-package com.victoriametrics.metrics;
+package com.victoriametrics.client.metrics;
 
 import java.util.concurrent.atomic.AtomicLong;
