+++ src/main/java/com/victoriametrics/client/metrics/Gauge.java
@@ -1,4 +1,4 @@
-package com.victoriametrics.metrics;
+package com.victoriametrics.client.metrics;
 
 import java.util.Objects;
 import java.util.function.Supplier;
