+++ src/main/java/com/victoriametrics/client/utils/Pair.java
@@ -1,4 +1,4 @@
-package com.victoriametrics.utils;
+package com.victoriametrics.client.utils;
 
 public class Pair<K, V> {
