+++ src/test/java/com/victoriametrics/metrics/client/MetricCollectionTest.java
@@ -1,15 +1,19 @@
-package com.victoriametrics.metrics;
+package com.victoriametrics.metrics.client;
 
+import com.victoriametrics.client.metrics.MetricCollection;
+import com.victoriametrics.client.metrics.Counter;
+import com.victoriametrics.client.metrics.Gauge;
+import com.victoriametrics.client.metrics.Histogram;
 import com.victoriametrics.validator.InvalidMetricNameException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class CollectionTest {
+public class MetricCollectionTest {
 
     @Test
     public void createCounterMetricWithBuilder() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         Counter counterNoLabels = collection.createCounter()
                 .name("foo")
                 .then()
@@ -28,7 +32,7 @@ public class CollectionTest {
 
     @Test
     public void createGaugeMetricWithBuilder() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         Gauge gaugeNoLabels = collection.createGauge()
                 .withSupplier(() -> 1.0)
                 .name("foo")
@@ -49,7 +53,7 @@ public class CollectionTest {
 
     @Test
     public void createHistogramMetricWithBuilder() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         Histogram histogramNoLabels = collection.createHistogram()
                 .name("foo")
                 .then()
@@ -68,7 +72,7 @@ public class CollectionTest {
 
     @Test
     public void createMetricsByName() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         Counter counter = collection.getOrCreateCounter("foo");
         Gauge gauge = collection.getOrCreateGauge("foo{bar=\"value1\"}", () -> 2.0);
         Histogram histogram = collection.getOrCreateHistogram("foo{bar=\"value1\", baz=\"value2\"}");
@@ -81,14 +85,14 @@ public class CollectionTest {
 
     @Test
     public void createMetricWithInvalidName_thenThrowException() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         assertThrowsExactly(InvalidMetricNameException.class, () -> collection.getOrCreateHistogram("foo{"));
         assertThrowsExactly(InvalidMetricNameException.class, () -> collection.getOrCreateHistogram("foo{label=}"));
     }
 
     @Test
     public void testGetOrCreateMetric() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         Counter counter1 = collection.getOrCreateCounter("foo");
         Counter counter2 = collection.getOrCreateCounter("foo");
         Counter counter3 = collection.getOrCreateCounter("foo");
@@ -109,7 +113,7 @@ public class CollectionTest {
 
     @Test
     public void tesGetMetricByName() {
-        Collection collection = Collection.create();
+        MetricCollection collection = MetricCollection.create();
         collection.getOrCreateCounter("foo");
         Counter counter = collection.getMetric("foo");
         assertNotNull(counter);
